FIX : Refactor. removed all entity from repository and moved into usecase.

+++ libs/application/use-cases/meeting/delete-meeting.usecase.ts
@@ -1,5 +1,6 @@
+import { Meeting } from "../../../domain/entities/meeting.entity";
 import { Messages } from "../../../shared/constants/messages";
-import { ForbiddenError } from "../../../shared/errors/app-error";
+import { ForbiddenError, NotFoundError } from "../../../shared/errors/app-error";
 import { IMeetingRepository } from "../../interfaces/meeting-repository.interface";
 
 export class DeleteMeetingUseCase {
@@ -8,7 +9,11 @@ export class DeleteMeetingUseCase {
   ) { }
 
   async execute({ id,orgId}:{id: string,orgId:string}) {
-    const meeting = await this.meetingRepo.findById(id);
+    const meetingDTO = await this.meetingRepo.findById(id);
+    if (!meetingDTO) {
+      throw new NotFoundError(Messages.MEETING_NOT_FOUND);
+    }
+    const meeting = new Meeting(meetingDTO);
      if (!meeting?.ownedBy(orgId)) throw new ForbiddenError(Messages.FORBIDDEN);
     await this.meetingRepo.delete(id);
   }
